requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.1
beautifulsoup4==4.6.0
boto3==1.12.39
botocore==1.15.39
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-dotenv==1.4.2
django-modelcluster==5.0.1
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
docutils==0.15.2
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.9
isort==4.3.21
jmespath==0.9.5
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.2
psycopg2==2.8.4
pudb==2019.2
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
urwid==2.1.0
wagtail==2.8
wagtailtrans==2.1
waitress==1.4.3
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
wrapt==1.11.2
